style: move rogue staking tests to dedicated file
