+++ test/staking/StakingProxy.Inflation.spec.ts
@@ -98,133 +98,6 @@ describe("Inflation", async () => {
             await expect(rigoToken.changeMintingAddress(user1.address)).to.be.reverted
             await expect(rigoToken.mintToken(AddressZero, 5)).to.be.reverted
         })
-
-        // this test should assure that a rogue upgrade of staking implementation won't affect token issuance
-        it('should revert on time anomalies', async () => {
-            const { inflation, stakingProxy, rigoToken } = await setupTests()
-            const StakingProxyInstance = await deployments.get("StakingProxy")
-            const StakingProxy = await hre.ethers.getContractFactory("StakingProxy")
-            const proxy = StakingProxy.attach(StakingProxyInstance.address)
-            // we must preserve storage in order to overwrite the correct storage slot
-            // unfortunately we must declare all preceding variables
-            const source = `
-            contract RogueStaking {
-                address public owner;
-                mapping(address => bool) public authorized;
-                address[] public authorities;
-                address public stakingContract;
-                mapping(uint8 => StoredBalance) internal _globalStakeByStatus;
-                mapping(uint8 => mapping(address => StoredBalance)) internal _ownerStakeByStatus;
-                mapping(address => mapping(bytes32 => StoredBalance)) internal _delegatedStakeToPoolByOwner;
-                mapping(bytes32 => StoredBalance) internal _delegatedStakeByPoolId;
-                mapping(address => bytes32) public poolIdByRbPoolAccount;
-                mapping(bytes32 => Pool) internal _poolById;
-                mapping(bytes32 => uint256) public rewardsByPoolId;
-                uint256 public currentEpoch;
-                uint256 public currentEpochStartTimeInSeconds;
-                mapping(bytes32 => mapping(uint256 => Fraction)) internal _cumulativeRewardsByPool;
-                mapping(bytes32 => uint256) internal _cumulativeRewardsByPoolLastStored;
-                mapping(address => bool) public validPops;
-                uint256 public epochDurationInSeconds;
-                uint32 public rewardDelegatedStakeWeight; // 1e15
-                uint256 public minimumPoolStake; // 1e19
-                uint32 public cobbDouglasAlphaNumerator; // 2
-                uint32 public cobbDouglasAlphaDenominator; // 3
-                address public inflation;
-                struct PoolStats { uint256 feesCollected; uint256 weightedStake; uint256 membersStake; }
-                struct AggregatedStats { uint256 rewardsAvailable; uint256 numPoolsToFinalize; uint256 totalFeesCollected; uint256 totalWeightedStake; uint256 totalRewardsFinalized; }
-                struct StoredBalance { uint64 currentEpoch; uint96 currentEpochBalance; uint96 nextEpochBalance; }
-                struct Pool { address operator; address stakingPal; uint32 operatorShare; uint32 stakingPalShare; }
-                struct Fraction { uint256 numerator; uint256 denominator; }
-                function init() public {}
-                function setDuration(uint256 _duration) public { epochDurationInSeconds = _duration; }
-                function setStaking(address _staking) public { stakingContract = _staking; }
-                function setInflation(address _inflation) public { inflation = _inflation; }
-                function endEpoch() public returns (uint256) {
-                    bytes4 selector = bytes4(keccak256(bytes("mintInflation()")));
-                    bytes memory encodedCall = abi.encodeWithSelector(selector);
-                    (bool success, bytes memory data) = inflation.call(encodedCall);
-                    if (!success) { revert(string(data)); } return uint256(bytes32(data));
-                }
-                function getInflation() public view returns (uint256) {
-                    bytes4 selector = bytes4(keccak256(bytes("getEpochInflation()")));
-                    bytes memory encodedCall = abi.encodeWithSelector(selector);
-                    ( , bytes memory data) = inflation.staticcall(encodedCall);
-                    return uint256(bytes32(data));
-                }
-                function getParams() external view returns (uint256, uint32, uint256, uint32, uint32) {
-                    return (epochDurationInSeconds, 1, 1, 1, 1);
-                }
-            }`
-            const rogueImplementation = await deployContract(user1, source)
-            const rogueProxy = rogueImplementation.attach(proxy.address)
-            await proxy.addAuthorizedAddress(user1.address)
-            await expect(
-                proxy.detachStakingContract()
-            ).to.emit(proxy, "StakingContractDetachedFromProxy")
-            await expect(stakingProxy.endEpoch()).to.be.revertedWith("STAKING_ADDRESS_NULL_ERROR")
-            await expect(
-                proxy.attachStakingContract(rogueImplementation.address)
-            ).to.be.emit(proxy, "StakingContractAttachedToProxy").withArgs(rogueImplementation.address)
-            await rogueProxy.setInflation(inflation.address)
-            // TODO: following tests work, but do not return expected errror
-            // max 90 days duration
-            await rogueProxy.setDuration(77760001)
-            await expect(rogueProxy.endEpoch()).to.be.revertedWith("INFLATION_TIME_ANOMALY_ERROR")
-            // min 5 days duration
-            await rogueProxy.setDuration(431999)
-            await expect(rogueProxy.endEpoch()).to.be.revertedWith("INFLATION_TIME_ANOMALY_ERROR")
-            await rogueProxy.setDuration(432000)
-            await expect(rogueProxy.endEpoch()).to.emit(rigoToken, "TokenMinted")
-            await expect(rogueProxy.endEpoch()).to.be.reverted
-            await timeTravel({ days: 5, mine:true })
-            const mintAmount = await rogueProxy.getInflation()
-            expect(await rogueProxy.callStatic.endEpoch()).to.be.eq(mintAmount)
-            await expect(rogueProxy.endEpoch()).to.emit(rigoToken, "TokenMinted").withArgs(proxy.address, mintAmount)
-            await expect(rogueProxy.endEpoch()).to.be.revertedWith("INFLATION_EPOCH_END_ERROR")
-        })
-
-        it('should not attach staking with invalid params', async () => {
-            const { inflation, stakingProxy, rigoToken } = await setupTests()
-            const StakingProxyInstance = await deployments.get("StakingProxy")
-            const StakingProxy = await hre.ethers.getContractFactory("StakingProxy")
-            const proxy = StakingProxy.attach(StakingProxyInstance.address)
-            const source = `
-            contract Staking {
-                event Init(address from);
-                uint256 public epochDurationInSeconds = 1 days;
-                uint256 public minimumPoolStake = 1e19;
-                uint32 public cobbDouglasAlphaNumerator = 2;
-                uint32 public cobbDouglasAlphaDenominator = 3;
-                uint32 public rewardDelegatedStakeWeight = 1e5;
-                uint32 public PPM_DENOMINATOR = 1e6;
-                address private inflation;
-                bytes4 immutable private SELECTOR = bytes4(keccak256(bytes("mintInflation()")));
-                function init() public { emit Init(msg.sender); }
-                function setTarget(address _inflation) external { inflation = _inflation; }
-                function endEpoch() external returns (uint256) {
-                    (bool success, bytes memory data) = inflation.call(abi.encodeWithSelector(SELECTOR));
-                }
-                function setDuration() external { epochDurationInSeconds = 0; }
-            }`
-            const mockImplementation = await deployContract(user1, source)
-            await mockImplementation.setTarget(inflation.address)
-            await proxy.addAuthorizedAddress(user1.address)
-            await expect(
-                proxy.detachStakingContract()
-            ).to.emit(proxy, "StakingContractDetachedFromProxy")
-            // staking contract should revert on adding contract with invalid parameters
-            /*await expect(
-                proxy.attachStakingContract(mockImplementation.address)
-            ).to.be.reverted
-            */
-            //await expect(stakingProxy.endEpoch()).to.be.revertedWith("STAKING_ADDRESS_NULL_ERROR")
-            /*const stakingInstance = await deployments.get("Staking")
-            await expect(
-                proxy.attachStakingContract(stakingInstance.address)
-            ).to.be.revertedWith("STAKING_SCHEDULER_ALREADY_INITIALIZED_ERROR")
-            */
-        })
     })
 
     describe("timeUntilNextClaim", async () => {
